Move query functionality to QueryFunction class
