+++ poolakey/src/main/java/com/phelat/poolakey/billing/query/QueryFunctionRequest.kt
@@ -1,0 +1,10 @@
+package com.phelat.poolakey.billing.query
+
+import com.phelat.poolakey.PurchaseType
+import com.phelat.poolakey.billing.FunctionRequest
+import com.phelat.poolakey.callback.PurchaseQueryCallback
+
+internal class QueryFunctionRequest(
+    val purchaseType: PurchaseType,
+    val callback: PurchaseQueryCallback.() -> Unit
+) : FunctionRequest
